pyproject.toml
[project]
authors = [
    {name = "Lucas Lu"},
]
requires-python = "<4.0,>=3.11"
dependencies = [
    "streamlit>=1.43.2",
    "pandas>=2.2.3",
    "faiss-cpu>=1.10.0",
    "matplotlib>=3.10.1",
    "pymysql>=1.1.1",
    "langgraph>=0.3.18",
    "google-genai>=1.7.0",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.9",
]
name = "college-information-llm"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false
